Add Scripting folder to separate script/binding code from the game/engine logic
