+++ src/Application/CMakeLists.txt
@@ -2,7 +2,6 @@ cmake_minimum_required(VERSION 3.27 FATAL_ERROR)
 
 set(APPLICATION_SOURCES
         Game.cpp
-        GameLuaBindings.cpp
         GameConfig.cpp
         GameKeyboardController.cpp
         GameMenu.cpp
@@ -15,7 +14,6 @@ set(APPLICATION_SOURCES
 
 set(APPLICATION_HEADERS
         Game.h
-        GameLuaBindings.h
         GameConfig.h
         GameKeyboardController.h
         GameMenu.h
@@ -25,8 +23,7 @@ set(APPLICATION_HEADERS
         GameStarterOptions.h
         GameTraceHandler.h
         GameWindowHandler.h
-        GameIocContainer.h
-	LuaItemQueryTable.h)
+        GameIocContainer.h)
 
 add_library(application STATIC ${APPLICATION_SOURCES} ${APPLICATION_HEADERS})
 target_check_style(application)
@@ -48,6 +45,7 @@ target_link_libraries(application
         library_platform_null
         library_platform_implementation
         library_environment_implementation
+        scripting
         utility
 	PRIVATE
 	libluajit
