+++ src/CMakeLists.txt
@@ -8,4 +8,5 @@ add_subdirectory(GUI)
 add_subdirectory(Io)
 add_subdirectory(Library)
 add_subdirectory(Media)
+add_subdirectory(Scripting)
 add_subdirectory(Utility)
