+++ src/Scripting/LuaItemQueryTable.h
